+++ apps/rush-lib/src/api/RushConfiguration.ts
@@ -100,18 +100,90 @@ export interface IRushRepositoryJson {
 }
 
 /**
- * Part of IRushConfigurationJson.
+ * This represents the available PNPM store options
+ * @public
  */
-export interface IPnpmOptionsJson {
+export type PnpmStoreOptions = 'local' | 'global';
+
+/**
+ * Options for the package manager.
+ * @public
+ */
+export interface IPackageManagerOptionsJsonBase {
+  /**
+   * Enviroment variables for the package manager
+   */
+  environmentVariables?: IConfigurationEnvironment
+}
+
+/**
+ * A collection of environment variables
+ * @public
+ */
+export interface IConfigurationEnvironment {
+  /**
+   * Environment variables
+   */
+  [environmentVariableName: string]: IConfigurationEnvironmentVariable;
+}
+
+/**
+ * Represents the value of an environment variable, and if the value should be overridden if the variable is set
+ * in the parent environment.
+ * @public
+ */
+export interface IConfigurationEnvironmentVariable {
+  /**
+   * Value of the environment variable
+   */
+  value: string;
+
+  /**
+   * Set to true to override the environment variable even if it is set in the parent environment.
+   * The default value is false.
+   */
+  override?: boolean;
+}
+
+/**
+ * Part of IRushConfigurationJson.
+ * @internal
+ */
+export interface INpmOptionsJson extends IPackageManagerOptionsJsonBase {
+}
+
+/**
+ * Part of IRushConfigurationJson.
+ * @internal
+ */
+export interface IPnpmOptionsJson extends IPackageManagerOptionsJsonBase {
+  /**
+   * The store resolution method for PNPM to use
+   */
+  pnpmStore?: PnpmStoreOptions;
+  /**
+   * Should PNPM fail if peer dependencies aren't installed?
+   */
   strictPeerDependencies?: boolean;
+  /**
+   * Defines the dependency resolution strategy PNPM will use
+   */
   resolutionStrategy?: ResolutionStrategy;
   preventManualShrinkwrapChanges?: boolean;
 }
 
 /**
  * Part of IRushConfigurationJson.
+ * @internal
  */
-export interface IYarnOptionsJson {
+export interface IYarnOptionsJson extends IPackageManagerOptionsJsonBase {
+  /**
+   * If true, then Rush will add the "--ignore-engines" option when invoking Yarn.
+   * This allows "rush install" to succeed if there are dependencies with engines defined in
+   * package.json which do not match the current environment.
+   *
+   * The default value is false.
+   */
   ignoreEngines?: boolean;
 }
 
@@ -144,6 +216,7 @@ export interface IRushConfigurationJson {
   projects: IRushConfigurationProjectJson[];
   eventHooks?: IEventHooksJson;
   hotfixChangeEnabled?: boolean;
+  npmOptions?: INpmOptionsJson;
   pnpmOptions?: IPnpmOptionsJson;
   yarnOptions?: IYarnOptionsJson;
   ensureConsistentVersions?: boolean;
@@ -167,6 +240,39 @@ export interface ICurrentVariantJson {
 }
 
 /**
+ * Options that all package managers share.
+ *
+ * @public
+ */
+export abstract class PackageManagerOptionsConfigurationBase implements IPackageManagerOptionsJsonBase {
+  /**
+   * Enviroment variables for the package manager
+   */
+  public readonly environmentVariables?: IConfigurationEnvironment
+
+  /** @internal */
+  protected constructor(json: IPackageManagerOptionsJsonBase) {
+    this.environmentVariables = json.environmentVariables;
+  }
+}
+
+/**
+ * Options that are only used when the NPM package manager is selected.
+ *
+ * @remarks
+ * It is valid to define these options in rush.json even if the NPM package manager
+ * is not being used.
+ *
+ * @public
+ */
+export class NpmOptionsConfiguration extends PackageManagerOptionsConfigurationBase {
+  /** @internal */
+  public constructor(json: INpmOptionsJson) {
+    super(json);
+  }
+}
+
+/**
  * Options that are only used when the PNPM package manager is selected.
  *
  * @remarks
@@ -175,7 +281,24 @@ export interface ICurrentVariantJson {
  *
  * @public
  */
-export class PnpmOptionsConfiguration {
+export class PnpmOptionsConfiguration extends PackageManagerOptionsConfigurationBase {
+  /**
+   * The method used to resolve the store used by PNPM.
+   *
+   * @remarks
+   * Available options:
+   *  - local: Use the standard Rush store path: common/temp/pnpm-store
+   *  - global: Use PNPM's global store path
+   */
+  public readonly pnpmStore: PnpmStoreOptions;
+
+  /**
+   * The path for PNPM to use as the store directory.
+   *
+   * Will be overridden by environment variable RUSH_PNPM_STORE_PATH
+   */
+  public readonly pnpmStorePath: string;
+
   /**
    * If true, then Rush will add the "--strict-peer-dependencies" option when invoking PNPM.
    *
@@ -225,7 +348,16 @@ export class PnpmOptionsConfiguration {
   public readonly preventManualShrinkwrapChanges: boolean;
 
   /** @internal */
-  public constructor(json: IPnpmOptionsJson) {
+  public constructor(json: IPnpmOptionsJson, commonTempFolder: string) {
+    super(json);
+    this.pnpmStore = json.pnpmStore || 'local';
+    if (EnvironmentConfiguration.pnpmStorePathOverride) {
+      this.pnpmStorePath = EnvironmentConfiguration.pnpmStorePathOverride;
+    } else if (this.pnpmStore === 'global') {
+      this.pnpmStorePath = '';
+    } else {
+      this.pnpmStorePath = path.resolve(path.join(commonTempFolder, 'pnpm-store'));
+    }
     this.strictPeerDependencies = !!json.strictPeerDependencies;
     this.resolutionStrategy = json.resolutionStrategy || 'fewer-dependencies';
     this.preventManualShrinkwrapChanges = !!json.preventManualShrinkwrapChanges
@@ -241,7 +373,7 @@ export class PnpmOptionsConfiguration {
  *
  * @public
  */
-export class YarnOptionsConfiguration {
+export class YarnOptionsConfiguration extends PackageManagerOptionsConfigurationBase {
   /**
    * If true, then Rush will add the "--ignore-engines" option when invoking Yarn.
    * This allows "rush install" to succeed if there are dependencies with engines defined in
@@ -253,6 +385,7 @@ export class YarnOptionsConfiguration {
 
   /** @internal */
   public constructor(json: IYarnOptionsJson) {
+    super(json);
     this.ignoreEngines = !!json.ignoreEngines;
   }
 }
@@ -298,7 +431,6 @@ export class RushConfiguration {
   private _packageManagerWrapper: PackageManager;
   private _npmCacheFolder: string;
   private _npmTmpFolder: string;
-  private _pnpmStoreFolder: string;
   private _yarnCacheFolder: string;
   private _shrinkwrapFilename: string;
   private _tempShrinkwrapFilename: string;
@@ -331,6 +463,7 @@ export class RushConfiguration {
   private _repositoryDefaultBranch: string;
   private _repositoryDefaultRemote: string;
 
+  private _npmOptions: NpmOptionsConfiguration;
   private _pnpmOptions: PnpmOptionsConfiguration;
   private _yarnOptions: YarnOptionsConfiguration;
 
@@ -383,7 +516,6 @@ export class RushConfiguration {
 
     this._npmCacheFolder = path.resolve(path.join(this._commonTempFolder, 'npm-cache'));
     this._npmTmpFolder = path.resolve(path.join(this._commonTempFolder, 'npm-tmp'));
-    this._pnpmStoreFolder = path.resolve(path.join(this._commonTempFolder, 'pnpm-store'));
     this._yarnCacheFolder = path.resolve(path.join(this._commonTempFolder, 'yarn-cache'));
 
     this._changesFolder = path.join(this._commonFolder, RushConstants.changeFilesFolderName);
@@ -401,7 +533,9 @@ export class RushConfiguration {
     );
     this._experimentsConfiguration = new ExperimentsConfiguration(experimentsConfigFile);
 
-    this._pnpmOptions = new PnpmOptionsConfiguration(rushConfigurationJson.pnpmOptions || {});
+    this._npmOptions = new NpmOptionsConfiguration(rushConfigurationJson.npmOptions || {});
+    this._pnpmOptions = new PnpmOptionsConfiguration(rushConfigurationJson.pnpmOptions || {},
+      this._commonTempFolder);
     this._yarnOptions = new YarnOptionsConfiguration(rushConfigurationJson.yarnOptions || {});
 
     // TODO: Add an actual "packageManager" field in rush.json
@@ -443,10 +577,10 @@ export class RushConfiguration {
     this._shrinkwrapFilename = this._packageManagerWrapper.shrinkwrapFilename;
 
     this._tempShrinkwrapFilename = path.join(
-        this._commonTempFolder, this._shrinkwrapFilename
+      this._commonTempFolder, this._shrinkwrapFilename
     );
     this._packageManagerToolFilename = path.resolve(path.join(
-        this._commonTempFolder, `${this.packageManager}-local`, 'node_modules', '.bin', `${this.packageManager}`
+      this._commonTempFolder, `${this.packageManager}-local`, 'node_modules', '.bin', `${this.packageManager}`
     ));
 
     /// From "C:\repo\common\temp\pnpm-lock.yaml" --> "C:\repo\common\temp\pnpm-lock-preinstall.yaml"
@@ -455,9 +589,9 @@ export class RushConfiguration {
       parsedPath.name + '-preinstall' + parsedPath.ext);
 
     RushConfiguration._validateCommonRushConfigFolder(
-        this._commonRushConfigFolder,
-        this.packageManager,
-        this._shrinkwrapFilename
+      this._commonRushConfigFolder,
+      this.packageManager,
+      this._shrinkwrapFilename
     );
 
     this._projectFolderMinDepth = rushConfigurationJson.projectFolderMinDepth !== undefined
@@ -622,11 +756,11 @@ export class RushConfiguration {
       if (semver.major(Rush.version) !== semver.major(expectedRushVersion)
         || semver.minor(Rush.version) !== semver.minor(expectedRushVersion)) {
 
-          // If the major/minor are different, then make sure it's an older version.
-          if (semver.lt(Rush.version, expectedRushVersion)) {
-            throw new Error(`Unable to load ${rushJsonBaseName} because its RushVersion is`
-              + ` ${rushConfigurationJson.rushVersion}, whereas @microsoft/rush-lib is version ${Rush.version}.`
-              + ` Consider upgrading the library.`);
+        // If the major/minor are different, then make sure it's an older version.
+        if (semver.lt(Rush.version, expectedRushVersion)) {
+          throw new Error(`Unable to load ${rushJsonBaseName} because its RushVersion is`
+            + ` ${rushConfigurationJson.rushVersion}, whereas @microsoft/rush-lib is version ${Rush.version}.`
+            + ` Consider upgrading the library.`);
         }
       }
     }
@@ -721,9 +855,9 @@ export class RushConfiguration {
    * recognized config files.
    */
   private static _validateCommonRushConfigFolder(
-      commonRushConfigFolder: string,
-      packageManager: PackageManagerName,
-      shrinkwrapFilename: string
+    commonRushConfigFolder: string,
+    packageManager: PackageManagerName,
+    shrinkwrapFilename: string
   ): void {
     if (!FileSystem.exists(commonRushConfigFolder)) {
       console.log(`Creating folder: ${commonRushConfigFolder}`);
@@ -872,15 +1006,6 @@ export class RushConfiguration {
    */
   public get npmTmpFolder(): string {
     return this._npmTmpFolder;
-  }
-
-  /**
-   * The local folder where PNPM stores a global installation for every installed package
-   *
-   * Example: `C:\MyRepo\common\temp\pnpm-store`
-   */
-  public get pnpmStoreFolder(): string {
-    return this._pnpmStoreFolder;
   }
 
   /**
@@ -1123,6 +1248,13 @@ export class RushConfiguration {
 
   public get projectsByName(): Map<string, RushConfigurationProject> {
     return this._projectsByName;
+  }
+
+  /**
+   * {@inheritDoc NpmOptionsConfiguration}
+   */
+  public get npmOptions(): NpmOptionsConfiguration {
+    return this._npmOptions;
   }
 
   /**