+++ apps/rush-lib/src/cli/actions/InstallAction.ts
@@ -34,7 +34,10 @@ export class InstallAction extends BaseInstallAction {
       recheckShrinkwrap: false,
       networkConcurrency: this._networkConcurrencyParameter.value,
       collectLogFile: this._debugPackageManagerParameter.value!,
-      variant: this._variant.value
+      variant: this._variant.value,
+      // Because the 'defautltValue' option on the _maxInstallAttempts parameter is set,
+      // it is safe to assume that the value is not null
+      maxInstallAttempts: this._maxInstallAttempts.value!
     };
   }
 }