+++ apps/rush-lib/src/logic/InstallManager.ts
@@ -22,7 +22,8 @@ import {
   FileSystem,
   FileConstants,
   Sort,
-  PosixModeBits
+  PosixModeBits,
+  JsonObject
 } from '@microsoft/node-core-library';
 
 import { ApprovedPackagesChecker } from '../logic/ApprovedPackagesChecker';
@@ -35,7 +36,7 @@ import { Git } from '../logic/Git';
 import { LastInstallFlag } from '../api/LastInstallFlag';
 import { LinkManagerFactory } from '../logic/LinkManagerFactory';
 import { PurgeManager } from './PurgeManager';
-import { RushConfiguration, ICurrentVariantJson } from '../api/RushConfiguration';
+import { RushConfiguration, ICurrentVariantJson, IConfigurationEnvironment } from '../api/RushConfiguration';
 import { RushConfigurationProject } from '../api/RushConfigurationProject';
 import { RushConstants } from '../logic/RushConstants';
 import { ShrinkwrapFileFactory } from '../logic/ShrinkwrapFileFactory';
@@ -50,8 +51,6 @@ import { ShrinkwrapFilePolicy } from './policy/ShrinkwrapFilePolicy';
 // The PosixModeBits are intended to be used with bitwise operations.
 /* eslint-disable no-bitwise */
 
-const MAX_INSTALL_ATTEMPTS: number = 2;
-
 /**
  * The "noMtime" flag is new in tar@4.4.1 and not available yet for \@types/tar.
  * As a temporary workaround, augment the type.
@@ -61,6 +60,7 @@ import { RushGlobalFolder } from '../api/RushGlobalFolder';
 import { PackageManagerName } from '../api/packageManager/PackageManager';
 import { PnpmPackageManager } from '../api/packageManager/PnpmPackageManager';
 import { DependencySpecifier } from './DependencySpecifier';
+import { EnvironmentConfiguration } from '../api/EnvironmentConfiguration';
 
 // eslint-disable-next-line @typescript-eslint/interface-name-prefix
 export interface CreateOptions {
@@ -120,6 +120,11 @@ export interface IInstallManagerOptions {
    * The variant to consider when performing installations and validating shrinkwrap updates.
    */
   variant?: string | undefined;
+
+  /**
+   * Retry the install the specified number of times
+   */
+  maxInstallAttempts: number
 }
 
 /**
@@ -144,11 +149,17 @@ export class InstallManager {
     this._commonTempFolderRecycler = purgeManager.commonTempFolderRecycler;
     this._options = options;
 
-    this._commonNodeModulesMarker = new LastInstallFlag(this._rushConfiguration.commonTempFolder, {
+    const lastInstallState: JsonObject = {
       node: process.versions.node,
       packageManager: rushConfiguration.packageManager,
       packageManagerVersion: rushConfiguration.packageManagerToolVersion
-    });
+    }
+
+    if (lastInstallState.packageManager === 'pnpm') {
+      lastInstallState.storePath = rushConfiguration.pnpmOptions.pnpmStorePath;
+    }
+
+    this._commonNodeModulesMarker = new LastInstallFlag(this._rushConfiguration.commonTempFolder, lastInstallState);
   }
 
   /**
@@ -425,7 +436,7 @@ export class InstallManager {
           packageName: packageManager,
           version: this._rushConfiguration.packageManagerToolVersion,
           tempPackageTitle: `${packageManager}-local-install`,
-          maxInstallAttempts: MAX_INSTALL_ATTEMPTS,
+          maxInstallAttempts: this._options.maxInstallAttempts,
           // This is using a local configuration to install a package in a shared global location.
           // Generally that's a bad practice, but in this case if we can successfully install
           // the package at all, we can reasonably assume it's good for all the repositories.
@@ -817,7 +828,7 @@ export class InstallManager {
         'node_modules');
 
       // This marker file indicates that the last "rush install" completed successfully
-      const markerFileExistedAndWasValidAtStart: boolean = this._commonNodeModulesMarker.isValid();
+      const markerFileExistedAndWasValidAtStart: boolean = this._commonNodeModulesMarker.checkValidAndReportStoreIssues();
 
       // If "--clean" or "--full-clean" was specified, or if the last install was interrupted,
       // then we will need to delete the node_modules folder.  Otherwise, we can do an incremental
@@ -898,6 +909,38 @@ export class InstallManager {
           // Example: "C:\MyRepo\common\temp\npm-local\node_modules\.bin\npm"
           const packageManagerFilename: string = this._rushConfiguration.packageManagerToolFilename;
 
+          let packageManagerEnv: NodeJS.ProcessEnv = process.env;
+
+          let configurationEnvironment: IConfigurationEnvironment | undefined = undefined;
+
+          if (this._rushConfiguration.packageManager === 'npm') {
+            if (
+              this._rushConfiguration.npmOptions &&
+              this._rushConfiguration.npmOptions.environmentVariables
+            ) {
+              configurationEnvironment = this._rushConfiguration.npmOptions.environmentVariables;
+            }
+          } else if (this._rushConfiguration.packageManager === 'pnpm') {
+            if (
+              this._rushConfiguration.pnpmOptions &&
+              this._rushConfiguration.pnpmOptions.environmentVariables
+            ) {
+              configurationEnvironment = this._rushConfiguration.pnpmOptions.environmentVariables;
+            }
+          } else if (this._rushConfiguration.packageManager === 'yarn') {
+            if (
+              this._rushConfiguration.yarnOptions &&
+              this._rushConfiguration.yarnOptions.environmentVariables
+            ) {
+              configurationEnvironment = this._rushConfiguration.yarnOptions.environmentVariables;
+            }
+          }
+
+          packageManagerEnv = this._mergeEnvironmentVariables(
+            process.env,
+            configurationEnvironment
+          );
+
           // Is there an existing "node_modules" folder to consider?
           if (FileSystem.exists(commonNodeModulesFolder)) {
             // Should we delete the entire "node_modules" folder?
@@ -920,8 +963,8 @@ export class InstallManager {
                 const args: string[] = ['prune'];
                 this._pushConfigurationArgs(args, options);
 
-                Utilities.executeCommandWithRetry(MAX_INSTALL_ATTEMPTS, packageManagerFilename, args,
-                  this._rushConfiguration.commonTempFolder);
+                Utilities.executeCommandWithRetry(this._options.maxInstallAttempts, packageManagerFilename, args,
+                  this._rushConfiguration.commonTempFolder, packageManagerEnv);
 
                 // Delete the (installed image of) the temp projects, since "npm install" does not
                 // detect changes for "file:./" references.
@@ -969,10 +1012,10 @@ export class InstallManager {
           }
 
           try {
-            Utilities.executeCommandWithRetry(MAX_INSTALL_ATTEMPTS, packageManagerFilename,
+            Utilities.executeCommandWithRetry(this._options.maxInstallAttempts, packageManagerFilename,
               installArgs,
               this._rushConfiguration.commonTempFolder,
-              undefined,
+              packageManagerEnv,
               false, () => {
                 if (this._rushConfiguration.packageManager === 'pnpm') {
                   console.log(colors.yellow(`Deleting the "node_modules" folder`));
@@ -984,16 +1027,19 @@ export class InstallManager {
 
                   Utilities.createFolderWithRetry(commonNodeModulesFolder);
                 }
-            });
+              });
           } catch (error) {
             // All the install attempts failed.
 
-            if (this._rushConfiguration.packageManager === 'pnpm') {
+            if (
+              this._rushConfiguration.packageManager === 'pnpm' &&
+              this._rushConfiguration.pnpmOptions.pnpmStore === 'local'
+            ) {
               // If the installation has failed even after the retries, then pnpm store may
               // have got into a corrupted, irrecoverable state. Delete the store so that a
               // future install can create the store afresh.
               console.log(colors.yellow(`Deleting the "pnpm-store" folder`));
-              this._commonTempFolderRecycler.moveFolder(this._rushConfiguration.pnpmStoreFolder);
+              this._commonTempFolderRecycler.moveFolder(this._rushConfiguration.pnpmOptions.pnpmStorePath);
             }
 
             throw error;
@@ -1027,6 +1073,48 @@ export class InstallManager {
           console.log('');
         });
     });
+  }
+
+  private _mergeEnvironmentVariables(
+    baseEnv: NodeJS.ProcessEnv,
+    environmentVariables?: IConfigurationEnvironment
+  ): NodeJS.ProcessEnv {
+    const packageManagerEnv: NodeJS.ProcessEnv = baseEnv;
+
+    if (environmentVariables) {
+      // eslint-disable-next-line guard-for-in
+      for (const envVar in environmentVariables) {
+        let setEnvironmentVariable: boolean = true;
+        console.log(`\nProcessing definition for environment variable: ${envVar}`);
+
+        if (baseEnv.hasOwnProperty(envVar)) {
+          setEnvironmentVariable = false;
+          console.log(`Environment variable already defined:`);
+          console.log(`  Name: ${envVar}`);
+          console.log(`  Existing value: ${baseEnv[envVar]}`);
+          console.log(`  Value set in rush.json: ${environmentVariables[envVar].value}`);
+
+          if (environmentVariables[envVar].override) {
+            setEnvironmentVariable = true;
+            console.log(`Overriding the environment variable with the value set in rush.json.`);
+          }
+          else {
+            console.log(colors.yellow(`WARNING: Not overriding the value of the environment variable.`));
+          }
+        }
+
+        if (setEnvironmentVariable) {
+          if (this._options.debug) {
+            console.log(`Setting environment variable for package manager.`);
+            console.log(`  Name: ${envVar}`);
+            console.log(`  Value: ${environmentVariables[envVar].value}`);
+          }
+          packageManagerEnv[envVar] = environmentVariables[envVar].value;
+        }
+      }
+    }
+
+    return packageManagerEnv;
   }
 
   private _checkIfReleaseIsPublished(): Promise<boolean> {
@@ -1175,7 +1263,15 @@ export class InstallManager {
         args.push('--verbose');
       }
     } else if (this._rushConfiguration.packageManager === 'pnpm') {
-      args.push('--store', this._rushConfiguration.pnpmStoreFolder);
+      // Only explicitly define the store path if `pnpmStore` is using the default, or has been set to
+      // 'local'.  If `pnpmStore` = 'global', then allow PNPM to use the system's default
+      // path.  In all cases, this will be overridden by RUSH_PNPM_STORE_PATH
+      if (
+        this._rushConfiguration.pnpmOptions.pnpmStore === 'local' ||
+        EnvironmentConfiguration.pnpmStorePathOverride
+      ) {
+        args.push('--store', this._rushConfiguration.pnpmOptions.pnpmStorePath);
+      }
 
       // we are using the --no-lock flag for now, which unfortunately prints a warning, but should be OK
       // since rush already has its own install lock file which will invalidate the cache for us.